feat: add common layout and routing components, and demo for shared UI components
- Introduced CommonLayout component for consistent layout structure with navigation.
- Added CommonRoutes component to handle dynamic routing.
- Created ComponentsDemo to showcase usage of Button, Card, and Input components.
- Included NxWelcome component as a starter template for new projects.
- Updated index.ts to export new components for easier access.
