+++ libs/shared/ui/src/index.ts
@@ -2,3 +2,7 @@ export * from './lib/ui';
 export * from './lib/button';
 export * from './lib/card';
 export * from './lib/input';
+export * from './lib/common-layout';
+export * from './lib/common-routes';
+export * from './lib/nx-welcome';
+export * from './lib/components-demo';
